cleaned up show
